requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
colorclass==2.2.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.1
django-activity-stream==0.9.0
django-allauth==0.42.0
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-bootstrap-form==3.4
django-clever-selects==0.8.2
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-rest-auth==0.9.5
django-rest-authtoken==1.2.4
django-session-security==2.6.6
django-threadedcomments==1.2
djangorestframework==3.11.1
docopt==0.6.2
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
isort==5.4.2
jsonfield==3.1.0
kombu
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
#Pillow==7.2.0
pinax-notifications==6.0.0
pinax-templates==2.0.6
pip-upgrader==1.4.15
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-http-client==3.3.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
django-utils-six
#sendgrid
#sendgrid-django
#sendgrid==3.6.5
#sendgrid-django==4.2.0
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
stream-framework==1.4.0
swapper==1.1.2.post1
terminaltables==3.1.0
toml==0.10.1
twilio==6.45.0
typed-ast==1.4.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
zipp==3.1.0
django-logentry-admin
django-otp==1.0.2
qrcode==6.1
django-colorfield
pymysql
mysqlclient
### email
#python-http-client==3.2.1
sendgrid==3.6.5
sendgrid-django==4.2.0
xlwt